invoice-processor/requirements.txt
pika==1.2.0
requests==2.26.0
mysql-connector-python==8.0.27
python-dotenv==0.19.2
